+++ mitra_activitypub/src/authentication.rs
@@ -134,6 +134,24 @@ pub enum AuthenticationError {
     InvalidPortableObject(#[from] PortableObjectAuthenticationError),
 }
 
+impl AuthenticationError {
+    pub fn ignore_if_missing_or_unsupported(self) -> Result<(), Self> {
+        match self {
+            // These errors can be ignored
+            AuthenticationError::NoHttpSignature
+                | AuthenticationError::NoJsonSignature => (),
+            // These errors should be warned about
+            AuthenticationError::UnsupportedVerificationMethod
+                | AuthenticationError::UnsupportedSignatureAlgorithm =>
+            {
+                log::warn!("{self}");
+            },
+            _ => return Err(self),
+        };
+        Ok(())
+    }
+}
+
 async fn get_signer(
     ap_client: &ApClient,
     db_pool: &DatabaseConnectionPool,