+++ mitra_activitypub/src/handlers/announce.rs
@@ -35,7 +35,7 @@ use mitra_validators::{
 };
 
 use crate::{
-    authentication::{verify_signed_object, AuthenticationError},
+    authentication::verify_signed_object,
     authority::Authority,
     identifiers::{
         canonicalize_id,
@@ -210,7 +210,8 @@ async fn handle_fep_1b12_announce(
         false, // fetch signer
     ).await {
         Ok(_) => activity.clone(),
-        Err(AuthenticationError::NoJsonSignature) => {
+        Err(error) => {
+            error.ignore_if_missing_or_unsupported()?;
             if is_same_origin(&announce_id, activity_id)? {
                 // Embedded activity can be trusted; don't fetch
                 // NOTE: assuming remote server validates C2S activities
@@ -230,7 +231,6 @@ async fn handle_fep_1b12_announce(
                 }
             }
         },
-        Err(other_error) => return Err(other_error.into()),
     };
     // Authorization
     verify_activity_owner(&activity)?;