+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Improved error messages when parsing verification methods.
 - Accept unsigned request to inbox endpoint if activity is signed.
 - Improved error message when verification of FEP-1b12 and FEP-171b integrity proof fails.
+- Ignore HTTP and JSON signatures where unsupported algorithms are used.
 
 ### Fixed